CXFunction.go: Add a selector that we forgot when creating the file.
Also: Delete empty files: getters.go, removers.go and selectors.go
