--- cx/getters.go
@@ -1,7 +1,0 @@
-package base
-
-import (
-	"errors"
-	"fmt"
-)
-
