--- cx/selectors.go
@@ -1,35 +1,0 @@
-package base
-
-import (
-	"errors"
-	"fmt"
-)
-
-func (fn *CXFunction) SelectExpression(line int) (*CXExpression, error) {
-	// prgrmStep := &CXProgramStep{
-	// 	Action: func(cxt *CXProgram) {
-	// 		if mod, err := cxt.GetCurrentPackage(); err == nil {
-	// 			if fn, err := mod.GetCurrentFunction(); err == nil {
-	// 				fn.SelectExpression(line)
-	// 			}
-	// 		}
-	// 	},
-	// }
-	// saveProgramStep(prgrmStep, fn.Context)
-	if len(fn.Expressions) == 0 {
-		return nil, errors.New("There are no expressions in this function")
-	}
-
-	if line >= len(fn.Expressions) {
-		line = len(fn.Expressions) - 1
-	}
-
-	if line < 0 {
-		line = 0
-	}
-
-	expr := fn.Expressions[line]
-	fn.CurrentExpression = expr
-
-	return expr, nil
-}
